refactor(domain): change CaloriesBurned data type from double to int

+++ FitnessTracker.Business/DTOs/CreateWorkoutRequest.cs
@@ -5,7 +5,7 @@ public record CreateWorkoutRequest
     public string Title { get; init; } = null!;
     public int TypeId { get; init; }
     public int DurationMinutes { get; init; }
-    public double CaloriesBurned { get; init; }
+    public int CaloriesBurned { get; init; }
     public DateTime WorkoutDate { get; init; }
     public List<CreateExerciseRequest> Exercises { get; init; } = [];
 };
